+++ CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#61](https://github.com/elgervb/chrome-readinglist2/issues/61)
+  Log analytics events
 
 ## 0.3.0
-
 ### Added
 - [#39](https://github.com/elgervb/chrome-readinglist2/issues/39)
   Link to extension reviews in Chrome store